fixerror

+++ serverdb/src/controller/Admin/DeleteUserBYID.ts
@@ -6,7 +6,15 @@ export const DeleteUserByID = async (req: Request , res: Response ) => {
     try {
         const { id } = req.params;
         const client = await dbConnect();
-        await client.query(`DELETE FROM User WHERE _id = ?` ,  id )
+        const check:any = await client.query(`SELECT * FROM User WHERE _id = ?`, id)
+        
+        if (check[0].length < 1) {
+            res.status(404).send({
+                message: "User not found"
+            })
+            return false;
+        }
+        await client.query(`DELETE FROM User WHERE _id = ?` , id)
         return res.status(200).send({
             message: "Delete user successed",
         })
